From Anders F Bjorklund:
The current definition of int32_t, uint32_t, uintptr_t
doesn't match what the Mac system headers already have...

Since it's an old 16/32 bit platform, the system headers
expect the types to use "long" and not "int" like they do.

+++ include/SDL_config_macos.h
@@ -31,9 +31,9 @@ typedef signed char	int8_t;
 typedef unsigned char	uint8_t;
 typedef signed short	int16_t;
 typedef unsigned short	uint16_t;
-typedef signed int	int32_t;
-typedef unsigned int	uint32_t;
-typedef unsigned int	uintptr_t;
+typedef signed long	int32_t;
+typedef unsigned long	uint32_t;
+typedef unsigned long	uintptr_t;
 
 /* Useful headers */
 #define HAVE_STDIO_H	1
